Move Cirò Marina GeoJSON to src/data and update imports
